pyproject.toml
[project]
name = "fast-zero"
version = "0.1.0"
description = ""
authors = [{ name = "Desenv2 - (Guilherme)", email = "hexadesenv2@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = ["fastapi[standard] (>=0.115.12,<0.116.0)", "sqlalchemy (>=2.0.47,<3.0.0)", "pydantic-settings (>=2.13.1,<3.0.0)", "alembic (>=1.18.4,<2.0.0)", "pwdlib[argon2] (>=0.3.0,<0.4.0)", "pyjwt (>=2.11.0,<3.0.0)", "tzdata (>=2025.3,<2026.0)"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.14.11,<0.15.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
]

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = ['.']
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check'

pre_format = 'ruff check --fix'
format = 'ruff format'

run = 'fastapi dev fast_zero/app.py'

pre_test = 'task lint'
test = "pytest -s -x --cov=fast_zero -vv"
post_test = 'coverage html'

database = 'pipx run harlequin database.db'
